% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Library.r
\name{SPACox.one.SNP}
\alias{SPACox.one.SNP}
\title{SaddlePoint Approximation implementation of Cox regression surival analysis (One-SNP-version)}
\usage{
SPACox.one.SNP(
  g,
  obj.null,
  Cutoff = 2,
  impute.method = "fixed",
  missing.cutoff = 0.15,
  min.maf = 1e-04,
  CovAdj.cutoff = 5e-05,
  G.model = "Add"
)
}
\arguments{
\item{g}{a numeric genotype vector. Missing genotype should be coded as NA. Both hard-called and imputed genotype data are supported.}

\item{obj.null}{an R object returned from function SPACox_Null_Model()}

\item{Cutoff}{a numeric value (Default: 2) to specify the standard deviation cutoff to be used.}

\item{impute.method}{a character string (default: "fixed") to specify the method to impute missing genotypes.}

\item{missing.cutoff}{a numeric value (default: 0.15) to specify the cutoff of the missing rates.}

\item{min.maf}{a numeric value (default: 0.0001) to specify the cutoff of the minimal MAF.}

\item{CovAdj.cutoff}{a numeric value (default: 5e-5). If the p-value is less than this cutoff, then we would use an additional technic to adjust for covariates.}

\item{G.model}{a character string (default: "Add") to specify the genetic model. Options are "Add", "Dom", and "Rec".}
}
\value{
the same as function SPACox.
}
\description{
One-SNP-version SPACox function. This function is to facilitate users that prefer reading and analyzing genotype line-by-line.
}
