% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Library.r
\name{SPACox.one.SNP}
\alias{SPACox.one.SNP}
\title{SaddlePoint Approximation implementation of Cox regression surival analysis (One-SNP-version)}
\usage{
SPACox.one.SNP(g, obj.null, Cutoff = 2, impute.method = "fixed",
  missing.cutoff = 0.15, min.maf = 1e-04, CovAdj.cutoff = 5e-05,
  G.model = "Add")
}
\arguments{
\item{g}{a numeric genotype vector. Missing genotype should be coded as NA. Both hard-called and imputed genotype data are supported.}

\item{others}{the same as function SPACox. NOTE that we do not check subject order in this one-snp-version !!!}
}
\value{
the same as function SPACox.
}
\description{
One-SNP-version SPACox function. This function is to facilitate users that prefer reading and analyzing genotype line-by-line.
}
